% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCMSQA-package.R
\docType{package}
\name{LCMSQA-package}
\alias{LCMSQA}
\alias{LCMSQA-package}
\title{LCMSQA: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jaehyun Joo \email{jaehyunjoo@outlook.com}

Authors:
\itemize{
  \item Blanca Himes \email{bhimes@pennmedicine.upenn.edu}
}

}
\keyword{internal}
