% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQA.R
\name{runQA}
\alias{runQA}
\title{Start LCMS QA application}
\usage{
runQA()
}
\arguments{
\item{BPPARAM}{A \link[BiocParallel:BiocParallelParam-class]{BiocParallel::BiocParallelParam} object}
}
\value{
This function normally does not return; interrupt R to stop the
application (usually by pressing Ctrl + C or ESC)
}
\description{
Start LCMS QA application
}
\examples{

\donttest{runQA()}

}
